refactor: remove all deprecated functions
